+++ packages/appcd-telemetry/CHANGELOG.md
@@ -2,6 +2,8 @@
 
  * BREAKING CHANGE: Event names are no longer prefixed with `appcd-`.
  * feat: Scrub potentially sensitive data in telemetry data and error messages.
+ * feat: Added ability to set a `hardwareId` instead of relying on the telemetry system to identify
+   a machine's unique identifier.
  * chore: Updated dependencies.
 
 # v4.0.0 (Jun 12, 2020)